requirements.txt
streamlit>=1.28.0
tqdm
scipy
scikit-learn
torch
torchvision
opencv-python
numpy
requests
click
pillow
matplotlib
sign-language-translator>=0.7.0,<0.8.0
# sign-language-translator==0.7.2  # Temporarily disabled due to Python 3.12 compatibility issues
deep_translator==1.11.4